+++ lib/src/models/models.dart
@@ -1,5 +1,0 @@
-export 'app_update_info_android.dart';
-export 'install_state_android.dart';
-export 'install_status_android.dart';
-export 'update_availability_android.dart';
-export 'update_result_android.dart';